// src/apps/contender/enums/contender.enum.ts
export enum CONTENDER_EVENT_TYPES {
  NEW_SEMINAR_REACTION = 'NEW_SEMINAR_REACTION',
  NEW_SEMINAR_UNREACTION = 'NEW_SEMINAR_UNREACTION',
  NEW_SEMINAR_CHANGED_REACTION = 'NEW_SEMINAR_CHANGED_REACTION',

  NEW_SEMINAR_COMMENT = 'NEW_SEMINAR_COMMENT',
  NEW_SEMINAR_COMMENT_REACTION = 'NEW_SEMINAR_COMMENT_REACTION',
  NEW_SEMINAR_COMMENT_UNREACTION = 'NEW_SEMINAR_UNREACTION',
  NEW_SEMINAR_COMMENT_CHANGED_REACTION = 'NEW_SEMINAR_CHANGED_REACTION',

  NEW_SEMINAR_COMMENT_REPLY = 'NEW_SEMINAR_COMMENT_REPLY',
  NEW_SEMINAR_COMMENT_REPLY_REACTION = 'NEW_SEMINAR_COMMENT_REPLY_REACTION',
  NEW_SEMINAR_COMMENT_REPLY_UNREACTION = 'NEW_SEMINAR_COMMENT_REPLY_UNREACTION',
  NEW_SEMINAR_COMMENT_REPLY_CHANGED_REACTION = 'NEW_SEMINAR_COMMENT_REPLY_CHANGED_REACTION',

  NEW_INTERVIEW_REACTION = 'NEW_INTERVIEW_REACTION',
  NEW_INTERVIEW_UNREACTION = 'NEW_INTERVIEW_UNREACTION',
  NEW_INTERVIEW_CHANGED_REACTION = 'NEW_INTERVIEW_CHANGED_REACTION',

  NEW_INTERVIEW_COMMENT = 'NEW_INTERVIEW_COMMENT',
  NEW_INTERVIEW_COMMENT_REACTION = 'NEW_INTERVIEW_COMMENT_REACTION',
  NEW_INTERVIEW_COMMENT_UNREACTION = 'NEW_INTERVIEW_COMMENT_UNREACTION',
  NEW_INTERVIEW_COMMENT_CHANGED_REACTION = 'NEW_INTERVIEW_COMMENT_CHANGED_REACTION',

  NEW_INTERVIEW_COMMENT_REPLY = 'NEW_INTERVIEW_COMMENT_REPLY',
  NEW_INTERVIEW_COMMENT_REPLY_REACTION = 'NEW_INTERVIEW_COMMENT_REPLY_REACTION',
  NEW_INTERVIEW_COMMENT_REPLY_UNREACTION = 'NEW_INTERVIEW_COMMENT_REPLY_UNREACTION',
  NEW_INTERVIEW_COMMENT_REPLY_CHANGED_REACTION = 'NEW_INTERVIEW_COMMENT_REPLY_CHANGED_REACTION',

  NEW_INTERVIEW_QUESTION_REACTION = 'NEW_INTERVIEW_QUESTION_REACTION',
  NEW_INTERVIEW_QUESTION_UNREACTION = 'NEW_INTERVIEW_QUESTION_UNREACTION',
  NEW_INTERVIEW_QUESTION_CHANGED_REACTION = 'NEW_INTERVIEW_QUESTION_CHANGED_REACTION',

  NEW_INTERVIEW_QUESTION_COMMENT = 'NEW_INTERVIEW_QUESTION_COMMENT',
  NEW_INTERVIEW_QUESTION_COMMENT_REACTION = 'NEW_INTERVIEW_QUESTION_COMMENT_REACTION',
  NEW_INTERVIEW_QUESTION_COMMENT_UNREACTION = 'NEW_INTERVIEW_QUESTION_COMMENT_UNREACTION',
  NEW_INTERVIEW_QUESTION_COMMENT_CHANGED_REACTION = 'NEW_INTERVIEW_QUESTION_COMMENT_CHANGED_REACTION',

  NEW_INTERVIEW_QUESTION_COMMENT_REPLY = 'NEW_INTERVIEW_QUESTION_COMMENT_REPLY',
  NEW_INTERVIEW_QUESTION_COMMENT_REPLY_REACTION = 'NEW_INTERVIEW_QUESTION_COMMENT_REPLY_REACTION',
  NEW_INTERVIEW_QUESTION_COMMENT_REPLY_UNREACTION = 'NEW_INTERVIEW_QUESTION_COMMENT_REPLY_UNREACTION',
  NEW_INTERVIEW_QUESTION_COMMENT_REPLY_CHANGED_REACTION = 'NEW_INTERVIEW_QUESTION_COMMENT_REPLY_CHANGED_REACTION',

  NEW_INTERVIEW_ANSWER_REACTION = 'NEW_INTERVIEW_ANSWER_REACTION',
  NEW_INTERVIEW_ANSWER_UNREACTION = 'NEW_INTERVIEW_ANSWER_UNREACTION',
  NEW_INTERVIEW_ANSWER_CHANGED_REACTION = 'NEW_INTERVIEW_ANSWER_CHANGED_REACTION',

  NEW_INTERVIEW_ANSWER_COMMENT = 'NEW_INTERVIEW_ANSWER_COMMENT',
  NEW_INTERVIEW_ANSWER_COMMENT_REACTION = 'NEW_INTERVIEW_ANSWER_COMMENT_REACTION',
  NEW_INTERVIEW_ANSWER_COMMENT_UNREACTION = 'NEW_INTERVIEW_ANSWER_COMMENT_UNREACTION',
  NEW_INTERVIEW_ANSWER_COMMENT_CHANGED_REACTION = 'NEW_INTERVIEW_ANSWER_COMMENT_CHANGED_REACTION',

  NEW_INTERVIEW_ANSWER_COMMENT_REPLY = 'NEW_INTERVIEW_ANSWER_COMMENT_REPLY',
  NEW_INTERVIEW_ANSWER_COMMENT_REPLY_REACTION = 'NEW_INTERVIEW_ANSWER_COMMENT_REPLY_REACTION',
  NEW_INTERVIEW_ANSWER_COMMENT_REPLY_UNREACTION = 'NEW_INTERVIEW_ANSWER_COMMENT_REPLY_UNREACTION',
  NEW_INTERVIEW_ANSWER_COMMENT_REPLY_CHANGED_REACTION = 'NEW_INTERVIEW_ANSWER_COMMENT_REPLY_CHANGED_REACTION',
}

export enum CONTENDER_NOTIFICATION_TARGET_TYPES {
  SEMINAR = 'SEMINAR',
  SEMINAR_COMMENT = 'SEMINAR_COMMENT',
  SEMINAR_COMMENT_REPLY = 'SEMINAR_COMMENT_REPLY',

  INTERVIEW = 'INTERVIEW',
  INTERVIEW_COMMENT = 'INTERVIEW_COMMENT',
  INTERVIEW_COMMENT_REPLY = 'INTERVIEW_COMMENT_REPLY',

  INTERVIEW_QUESTION = 'INTERVIEW_QUESTION',
  INTERVIEW_QUESTION_COMMENT = 'INTERVIEW_QUESTION_COMMENT',
  INTERVIEW_QUESTION_COMMENT_REPLY = 'INTERVIEW_QUESTION_COMMENT_REPLY',

  INTERVIEW_ANSWER = 'INTERVIEW_ANSWER',
  INTERVIEW_ANSWER_COMMENT = 'INTERVIEW_ANSWER_COMMENT',
  INTERVIEW_ANSWER_COMMENT_REPLY = 'INTERVIEW_ANSWER_COMMENT_REPLY',
}